chore: 파일 실행시 진입이슈 파일 경로 수정
