+++ app/(tabs)/_layout.tsx
@@ -1,0 +1,5 @@
+import { Tabs } from 'expo-router';
+
+export default function TabLayout() {
+    return <Tabs />;
+}
