+++ app/(tabs)/index.tsx
@@ -3,7 +3,7 @@ import { View, Text } from 'react-native';
 export default function HomeScreen() {
     return (
         <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-            <Text style={{ fontSize: 24 }}>홈 화면</Text>
+            <Text style={{ fontSize: 24 }}>탭 홈 화면!</Text>
         </View>
     );
 }
