--- app/(tabs)/layout.tsx
@@ -1,5 +1,0 @@
-import { Stack } from 'expo-router';
-
-export default function Layout() {
-    return <Stack />;
-}
