+++ app/auth/login.tsx
@@ -1,0 +1,9 @@
+import { View, Text } from "react-native";
+
+export default function LoginScreen() {
+    return (
+        <View>
+            <Text>Login</Text>
+        </View>
+    );
+}
