+++ app/onboarding/index.tsx
@@ -1,0 +1,9 @@
+import { View, Text } from "react-native";
+
+export default function OnboardingIndex() {
+    return (
+        <View>
+            <Text>Onboarding 홈 화면</Text>
+        </View>
+    );
+}
