+++ babel.config.js
@@ -1,4 +1,4 @@
 module.exports = {
     presets: ['babel-preset-expo'],
-    plugins: ['expo-router/babel', 'babel-plugin-styled-components'], // Expo Router가 정적 파일을 올바르게 처리하도록 설정
+    plugins: [],
 }
